+++ lib/event_bus_manager_getx.dart
@@ -1,5 +1,0 @@
-/// GetX [GetxController] lifecycle support for [EventBusManager].
-library;
-
-export 'event_bus_manager.dart';
-export 'src/event_bus_getx_subscriber_mixin.dart';